Adjust included data size on icmp errors
The RFC for icmpv6 specifies that an errant packet should be included
in the returned ICMP packet, and that it should include up to the amount
needed to fill the minimum MTU (1280 bytes) if possible. The current code
included the Link header in that calculation but the RFC is referring
to the IP MTU not the link MTU. Some conformance tests check this and
report an error agains the stack for this.

The full header length should however continue to be used when allocating
header space.

Make the same change for IPv4 for consistency.

Fixes #5311

PiperOrigin-RevId: 352632116